This adds a new daemon (ap.mcp) and a new command-line utility (ap-ctl).
mcp will start, stop, and restart daemons and will return status.  Currently you
can stop/start/restart a single daemon or all of them at once.  You can only get
status for a single daemon.

Daemons are described in a new mcp.json file.  This gives the name of the
daemon, the binary to run, along with any command line parameters.  You can
indicate whether a daemon should be run as root or not.

In the JSON file, each daemon may also identify a single other daemon that it
depends on.  This dependency is fairly loosely enforced right now.  If you ask
mcp to launch multiple daemons, it will respect those dependencies just among
those daemons.  If a daemon has a dependency on one that is not in the current
set being launched, mcp will behave as if that daemon is already running; it
will not automatically launch that daemon as well.

Each daemon will open a new ZMQ connection to the mcp, which it can use to send
status updates.  Currently a daemon's state is a simple string.  mcp only
recognizes the states 'offline', 'online', and 'broken' as meaningful, but will
allow you to set any state you like.  This needs to get turned into a real state
machine at some point, but this seems to be working for now.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D16