Assorted network robustness changes
Summary:
Rather than declaring your preferred network device as a configuration property,
networkd will now select one for you.  It will prefer a wireless NIC that
supports VLANs, but will fall back to the built-in NIC if that's the only one
available.

The SSID/passphrase is now a first-level network property, rather than being a
property of the NIC.

When possible, use SIGHUP to trigger a hostapd reset instead of a SIGINT.

networkd will signal and wait for hostapd prior to exiting.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D23