Add support for wired clients
Summary:

We inventory all the network devices when ap.networkd starts up.  The
built-in ethernet port is used to connect our network to the WAN.  All
other wired ports (presumably USB adapters) are connected to a software
bridge, and are configured as NATted interfaces for clients.

There is a new 'wired' class, which is used for all clients connected to
a wired interface.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D25