Add support for an open 'captive portal' interface
Summary:

If the wireless hardware supports it, this will create an unprotected
network which presents a captive portal.  If the regularly configured
SSID is "foo", the open SSID will be called "foo-connect".  Any client
can connect to the open network, but the client will only be able to
talk to our web server.  This gives us a path into these devices to
perform our "hands-off autoconfiguration".

This wad introduces a new "unauthorized" class with a corresponding
subnet.  Any client connecting to foo-connect will be placed into that
class, even if it has previously been recognized as belonging to a
different class.

All http (ports 80 and 8000) packets on that subnet will be forwarded to
our webserver.  DNS traffic will be allowed through, as that may be
useful for our classification and seems to be required for Apple's (at
least) captive portal recognition to work correctly.  All remaining
packets will be dropped.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D28