ios profile generation
Summary:

If an ios device hits our connect network, have brightgate.net generate
a wireless profile.

Now that using the '-connect' network can automatically get you access
to the real network, it's not enabled by default.  You have to create a
@/network/connectssid property.  Until we have some sort of
authentication and/or approval mechanism associated with this, this
functionality should only be enabled for test or demo purposes.

Test Plan: Manual on my phone and Chris's.

Differential Revision: https://ph0.b10e.net/D35