T52 ap.dhcp4d should prioritize static assignments over client requests Factor out Device commonality
Summary:

The dhcp daemon wasn't detecting configd changes that defined static IP
assignments.  This meant that a static assignment could be overridden by
a client attempting to renew an old address.

'Device' was defined in two different places so the database code could
be built without pulling in the rest of our codebase.  The identifier
training code wasn't using the device model, so it wasn't able to report
identities in a user friendly way.

With this change, we have a new ap_common/device module that can be
included by our daemons and by standlone code.  model-train will load
the device database to provide more useful information to users.

I've also added default values for the different input files for the
model training code.  In many cases, the user now only has to provide a
path to the observations dump.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D65