First cut of Framework7+VueJS based single-page-application
Summary:
This is more or less a skeleton, with no flesh on the bones.  Out of a
feeling of caution, I've not yet hooked up the "client-web" target to
the global "make all" or "make install" targets.  For now, you have to
"make client-web" to get the application populated to your proto area.

Test Plan: Ongoing, not much yet.

Reviewers: stephen, cat, sw-dev

Differential Revision: https://ph0.b10e.net/D82