Centralize mock device data, rework data representation for more flexibility
Summary:
A step on the way to a new schema.  Also this opens the data up so
that we can index into it by device name from the "device details" page.

Test Plan:
Clicked all the buttons; things still seem to work.  Saw that details
page matches devices more closely.

Reviewers: stephen, cat

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D96