Handle expiring ring properties
Summary:
Currently we assume that a client's ring property can change, but not expire.
Until we require clients to use WPA2 Enterprise to join our network, we need a
way to effectively revoke their access after a period of time.  Moving those
clients back to the unenrolled ring is the simplest way to accomplish that.  An
alternate approach would be to explicitly reject clients using hostapd, but
maintaining an ever-growing list of blacklisted clients seems impractical.

Test Plan:
Assign my laptop to 'standard' with an expiration of 1 minutes.  After that
minute elapses, verify that my connection to the network has been severed and
that an attempt to reacquire the DHCP lease results in getting an 'unenrolled'
IP address.

Differential Revision: https://ph0.b10e.net/D100