switch cloud_rpc to protobufs2 for consistency
Summary:
Since GRPC seems to support protobufs2 and protobufs3, and
because we have serious doubts about protobufs3, we've elected to
converge for now on protobufs2.

Reviewers, Please help me mark as many fields in the RPCs "optional" as possible.

Test Plan: Ran on pi according to Jeff's notes.  Ran upbeat and inventory.

Reviewers: jeff, stephen

Reviewed By: jeff, stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D115