Use bridges to support multiple NICs per ring
Summary:
The 'wired' ring has been removed.  The wired NIC is now attached to the
'standard' bridge.  This lets us have wired and wireless clients in the
same security ring and on the same subnet.  In the future, we will
support multiple wired NICs, and their ring assignments will be
configurable.

Wired clients will initially be assigned to 'unenrolled', just as new
wireless clients are.  Because the NIC belongs to the standard ring,
clients will need to be assigned to that ring in order to be assigned an
appropriate DHCP address.  This flow is a bit awkward, but should get
simpler once we have a proper enrollment interface.

Each ring (other than Setup) now has a bridge, rather than the prior mix
of bridges and raw interfaces.  This gets rid of the logical->physical
NIC indirection and the special casing around the N_WIFI and N_WIRED
nics.  It would be possible to have a setup bridge as well, but that's
less of a win, since everything involving the setup network would remain
a special case.

hostapd insists on controlling the bridges it uses for wireless VLANs.
Rather than modifying hostapd, networkd will play along.  When launching
hostapd, it will watch for the bridges to be created.  Once that
happens, it will attach any wired NICs to the appropriate bridge.
hostapd will assign any unrecognized clients to the 'default' interface,
which is not part of a VLAN.  Thus, networkd has to create the bridge
for unenrolled/non-vlan clients.

The @/interfaces part of the config tree has gone away.  The subnet
information that used to live there is now part of the per-ring
configuration.

Test Plan: Manual testing

Differential Revision: https://ph0.b10e.net/D119