identifierd gets too large
Summary:
This review addresses two issues.

The first is the 4MiB gRPC maximum message size. To work around this restriction
ap-rpc now limits the size of each inventory message to 2MiB. In addition, the
DeviceInfo message now contains an 'updated' field which ap-rpc will compare to
a manifest of when a DeviceInfo was last sent to the cloud. In the cloud
receiver we now create a directory for every client's DeviceInfo.

The second issue addressed is identifierd's tendency to consume too much memory.
The 'newData' entities structure was the culprit, trying to read and store the
device's lifetime of observations. Instead of storing all observation in core we
now stream them to disk every 15 minutes, writing only the DeviceInfo messages
that have been updated since they were last saved. To accomplish this I had to
decouple the inference data (testData) from the observation data (newData). The
former is a subset of the latter, but using the latter to reconsitute the former
was what caused the memory consumption problems. The inference data is now
written to disk as test_data.csv and the same frequency as the observation data.

Test Plan: Currently running on dogfood and svc0

Differential Revision: https://ph0.b10e.net/D133