Update model-train
Summary:
The functionality of model-train remains the same, but it now uses the device
database as its backend rather than maniplulating a collection of files. To
accomplish this, the bulk of device_db.go is now a package in the util/
directory. The device database still needs to be built by hand with
build_device_db.go.

model-train now reads DeviceInfo files and expects to be given a directory of
such files, which is the format used by the cloud when it receives inventory
reports.

Test Plan:
Used data from dogfood inventory reports to add new samples of existing devices
and new devices to the database.

Differential Revision: https://ph0.b10e.net/D145