Add support for 5GHz
Summary:
Allow channels and modes to be configured on a per-nic basis in
@/nodes/<nodeID>/<macID>.  Legal modes are 'g' and 'a' (which is
what hostapd calls 'ac').

Query the hardware for the list of supported channels (and, implicitly,
modes).  Ensure that the configured values are sane and supported.  If
they aren't, or if there are no configured options, choose a mode and
channel.  Currently, we will prefer 5GHz over 2.4GHz.  For 2.4GHz we
default to channel 6. For 5GHz, we choose a supported channel at random.

Test Plan:
Manual testing with a Panda N600 Dual Band 300Mbps Wireless N USB
Adapter

Differential Revision: https://ph0.b10e.net/D149