Make ap.relayd vet-clean Fix child-related crashes in ap.userauthd Set the wireless regulatory domain
Summary:

ap.userauthd was not clearing childProcess when the child died, which
was causing crashes in other threads.  There were also potential races
between the checking and use of childProcess, which were closed by
adding a mutex.

We weren't setting the regulatory domain, which limited the number of
channels we could use - and probably would have resulted in us using
illegal frequencies in many countries.  More immediately, it prevented
us from using the 5GHz frequencies.

Test Plan:

Removed letsencrypt certs, causing hostapd to crash.  Verified that
ap.userauthd was able to start, stop, and respond to config updates
cleanly.

Verified that my phone is still able to discover and operate a Wemo
outlet when the devices are in different rings.

Differential Revision: https://ph0.b10e.net/D157