Drop -mesh suffix on satellite APs
Summary:
During the initial development of the mesh/satellite support, it was
convenient to have the APs using different SSIDs.  That made it possible
to trigger a client migration quickly and deterministically between APs.
Now that we are beginning to deploy multi-node systems, it is more
desirable to provide a single SSID, allowing clients to migrate
automatically between APs.

Test Plan: Running on dogfood

Differential Revision: https://ph0.b10e.net/D163