Collect detailed network traffic statistics.
Summary:
Examines the headers of each of the captured packets to build a table of
per-device, per-endpoint traffic statistics.

Currently these statistics are collected in a single global map until watchd
exits.  The only way to extract the statistics is by looking at the output after
watchd exits:

Device                        Pkts Sent  Bytes Sent        Pkts Rcvd  Bytes Rcvd
6c:40:08:bf:63:26
       local:
                 192.168.133.1        1          91                1         138
                192.168.134.12     1276       87257             4840     7302953
      remote:
                  172.217.6.37        7         807                9        1363
               192.241.177.148        8        3642                1          74
                216.58.195.228        1          65                1          72
                 34.231.82.108        2         108                2         108
               173.194.203.189        1          65                1          72
                162.125.34.129        4        1305                4        1035
                  17.253.2.125        1          90                1          90
                 216.58.195.78        3         483                3         479
   aggregate:
                                   1304       93913             4863     7306384

Eventually these statistics will be rolled up into the cloud, with the local map
being trimmed.

ap-stats will be updated to extract these statistics from a running watchd, so
we can examine them without having to restart the daemon.

We currently use IP addresses for both local and remote endpoints.  For local
endpoints, we should be able to use the permanent device ID rather than the
transient IP address.  For inter-VLAN traffic, the mac address in the packet is
that of the router rather than the client, meaning we would have to do an
IP->Mac translation for each packet.  That can be done fairly easily, but is
left for a future wad.

Test Plan: manual

Differential Revision: https://ph0.b10e.net/D166