Make the default ring for new clients configurable
Summary:

Putting all new clients in the UNENROLLED ring is secure, but
requires more administrative intervention that might be desired.
This adds a new @/network/default_ring option, allowing the
administrator to give a higher level of access to any client with the
WPA password.  The most likely scenario is that new clients will be
given "guest ring" access rather than the isolated "unenrolled ring".

Note: this default only applies to clients that come in through the
front door (i.e., asking the DHCP server for an address).  Any client
we discover through nmap scans or other observed activity will still
be put in the UNENROLLED ring.

Differential Revision: https://ph0.b10e.net/D169