Modify mac address to accomodate WPA-EAP SSID
Summary:
When using multiple SSIDs, the wlan's mac address needs to have its
final octet set to 0.  This is already done to support a setup network,
but needs to be done for simultaneous WPA-PSK/WPA-EAP as well.

Test Plan: tested on Pi with USB WLAN mac address of 9e:ef:d5:fe:18:91

Differential Revision: https://ph0.b10e.net/D183