ap.mcp can collect and report on child processes' memory and CPU usage.
Summary:

This information can eventually be used to detect daemons that are
wildly out of scope, so they can be killed and restarted.  Currently the
information is available through ap-ctl:

nils@nilsap:~/dev/ctrl$ ap-ctl status all
      DAEMON      PID       RSS    SWAP    TIME        STATE    SINCE
  prometheus    26089     15 MB    0 MB     13s       online    Tue Jan 23 12:43:25
     brokerd    27024     10 MB    0 MB      1s       online    Tue Jan 23 12:50:18
     configd    26104     27 MB    0 MB      1s       online    Tue Jan 23 12:43:28
[...]
   userauthd    26131      4 MB    0 MB      0s       online    Tue Jan 23 12:43:28
      watchd     3292     52 MB    0 MB     54s       online    Tue Jan 23 15:17:56

aputil now has support for CPU and memory profiling.  This is plumbed
into watchd and identifierd (the two daemons most likely to have
issues), and can be triggered with a SIGUSR2.  This could also be wired
into the 'daemon out of scope' handling described above, so we can get a
snapshot of the daemon's state/behavior before killing it.  The profiles
can be uploaded to the cloud for offline analysis.

The various subcomponents of watchd are now launched/halted in parallel
rather than sequentially.

Test Plan: Manual

Differential Revision: https://ph0.b10e.net/D181