Remove false positives from IP blacklist
Summary:
The data in our hardcoded IP blacklist is stale and getting staler,
since we don't yet refresh is regularly.  (on the short-term "ToDo
list).  This pulls out 4 entries in there that used to be associated
with a botnet C&C server, but are now assigned to godoc instances.

This also removes a stray phishtank log message.

Test Plan: dogfood

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D186