Firewall bugfixes
Summary:
When we switched from using Sprintf to format our SQL insert to letting
the Exec() routine fill in the wildcards, we lost the automatic
translation of addresses and ports to strings.  That translation needs
to be done explicitly in the calling routine.

With the new, tighter firewall rules, the relayed UDP broadcast ports
need to be opened for internal clients.

Test Plan: Pi-based appliance with Pi and iPhone clients

Differential Revision: https://ph0.b10e.net/D202