Improve README.Product
Summary:
- Remove outdated information
- Reorganize
- Document packaging
- Document IoT credentials

(Note that there is parallel information about getting hostapd and tensorflow and golang in the wiki-- I'm not sure what to do about those just yet.  Let's consider this a step in the right direction.  I will also rename the file to README.Product.md but that messes up the codereview diff, so I'll do it in a separate step.)

Test Plan: Test builds and markdown rendering

Reviewers: stephen, nils, cat

Reviewed By: stephen, nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D206