T85: ap-configctl get clients could expand identifier ID to human readable
Summary:
"ap-configctl get clients" emits a column each for the device ID and the
confidence for that ID, but the ID is numeric, and neither that nor the
confidence is terribly meaningful to the end-user.

Instead, emit a single column for the device ID, using the name if the
confidence is better than 50%, and prefixed with a question mark if we're not
almost certain.

Test Plan: Manual testing with a few devices.

Reviewers: dp, stephen, nils

Reviewed By: dp, stephen, nils

Subscribers: buildbot

Maniphest Tasks: T85

Differential Revision: https://ph0.b10e.net/D213