Vulnerability detection and client quarantine
Summary:
This wad adds the ability to actively scan clients for vulnerabilities.  If any
are found we will send an event, which can eventually be turned into an alert
for the admin and/or the user.  If the vulnerability is considered threatening
enough, we will automatically quarantine the client as well.

ap-vuln-aggregate:

  This tool ingests a list of vulnerabilities to look for.  Given a host's IP
  address, it will probe that host for all of the listed vulnerabilities.  The
  results will be shown on the console, and optionally stored in a JSON file.

ap-inspect:

  This tool accepts an IP address and a CVE, and will test that host for the
  given CVE.  Currently, we are only capable of detecting a single CVE - and
  even that is not a particularly rigorous check.  While we may enhance the tool
  in the future, it currently exists primarily to demonstrate the generality of
  ap-vuln.

The watchd scanner has been enhanced to periodically scan for vulnerabilities,
as well as open ports.  It will invoke ap-vuln and consume the JSON data it
emits.

Test Plan:
manual testing with multiple Pis and with a wannacry-vulnerable
Windows VM

Differential Revision: https://ph0.b10e.net/D215