Fix IP expiration date formats
Summary:
I noticed in the output of `ap-configctl get clients` that the expiration date
of my phone was `2018-23-03`.  Several other places in the code made the same
mistake, so I corrected them all.

Test Plan:
Manually checked the output of `ap-configctl get clients` and `ap-configctl get @`,
and looked at `mcp.log` to find the one that `ap.configd` was writing.  Couldn't
find how to trigger the one in `ap.httpd`.

Reviewers: dp, nils

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D219