By default, 'get clients' will only show active clients
Summary:
A long-lived system like dogfood will collect a long list of systems
that connect rarely, or which have been retired.  A 'get clients' on
such a system will return an excessively large haystack, making it
difficult to find the desired needle(s).

With this change, 'get clients' will only show those clients that have
an IP address, and are thus likely to be of interest.  The full list is
still available with a '-a' option, or as raw data through 'get
@/clients'.

Reviewers: duvall

Differential Revision: https://ph0.b10e.net/D221