Remove env.sh
Summary:
It's a hassle to remember to source `env.sh` before building, or determine
whether you've already done it.  And our build scripts waste a lot of space
sourcing it over and over again.  What env.sh does can all be done in the
Makefile, so we might as well.

Test Plan:
Run `make packages` on a pi and on a cloud machine right after logging in, to
make sure that no extra environment variables need to be set.

Reviewers: stephen, dp, cat

Reviewed By: stephen, dp, cat

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D227