Property names must be unique
Summary:
Child nodes in the config tree are now represented as a name-indexed map, rather
than as a slice with each node containing a Name field.  This simplifes the code
that consumes and processes the config tree.  This also guarantees that any
given property path can match at most one property, removing potential
ambiguity.

This change in the config tree's structure causes a substantial change to the
config file's on-disk format, making it difficult to support upgrade older
config files.  To avoid rewriting all of the upgrade routines, we no longer
support upgrading files older than v10.

There is now a test module for ap.configd, making it possible to validate
changes to the daemon quickly and easily.  This new module obsoletes the old
apcfg test module.

Test Plan: running multi-node config with two Pis and an iphone client

Differential Revision: https://ph0.b10e.net/D233