Rationalize .gitignore files
Summary:
This should get rid of the persistently modified directories

    golang/src/bg/ap_common/iotcore/mocks/
    golang/src/bg/vendor/

and move ignore patterns closer to a place that better indicates
ownership.

Test Plan:
Run git status on a built workspace and see that it's clean.  Put a file in
`bg/bin` and see that it's not ignored.  Modify other files in `bg` (outside of
`bg/vendor`) and see that they're not ignored.  Run `git check-ignore` on a
handful of files that should be ignored and see that they're being excluded by
the expected rules.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D232