Add support for multi-step config operations
Summary:
The ConfigQuery message has been enhanced to allow the caller to provide
multiple operations that should be applied as a unit.

This initial implementation does not provide atomicity in the presence
of errors.  If a multi-step operation fails part way through, the
completed steps will not be rolled back.  That roll-back will come as a
follow-on to this work.

The ap.configd test framework now interacts with the main body of the
daemon code at the RPC invocation level, rather than by calling into
lower-level implementation routines.  This means 'go test' will exercise
more of the deployed code, and lays some necessary groundwork for
testing atomicity in the presence of errors.  Moving our interaction to
this higher level necessitated the changes around propfile (et.al.), so
we can prevent a 'go test' invocation from overwriting actual config
data with test data.

Test Plan:
ap.configd test support includes unit tests for the multi-operation support.

ap.dhcp4d and ap-userctl both make use of it.

Reviewers: stephen

Differential Revision: https://ph0.b10e.net/D237