T80 userauthd needs to handle configuration errors more gracefully
Summary:
In addition to fixes from Dan for consistently deriving the appliance FQDN, this changeset addressed T80, by testing certificates for existence and expiry (but not chain validity) and producing a self-signed certificate when a time-valid certificate does not exist.

When a new certificate is generated, an EventSysError is sent to sys.error; `ap.httpd` and `ap.userauthd` are listening to this topic, and will exit if the error indicates a new SSL certificate has been installed.

`ap-certcheck` is run in the nightly cronjob to cause new certificates to be requested, if they would expire in the next 25 hours.  This utility would be the location of any ACME invocation to update Let's Encrypt certificates.

Test Plan: `certificate_test.go` for testing self-signed generation.  Manual use of `ap-certcheck -force` to verify event sending and receipt.

Reviewers: dp, nils, cat, duvall

Reviewed By: dp, cat, duvall

Subscribers: buildbot

Tags: #sw-dev

Differential Revision: https://ph0.b10e.net/D228