Pull Raspberry Pi tensorflow from git instead of Phabricator Files
Summary:
The download from Files is broken for reasons that are not apparent,
making automated creation of the Raspberry Pi sysroot impossible.  The
ideal alternative would be to have the file checked in via git-lfs, but
that's not available on our development platforms.  This is an
alternative that gets the job done for now.

Test Plan:
Run build-multistrap-sysroot.sh by hand, make sure it puts a libtensorflow.so in
usr/local/lib with a hash that matches what's on my Raspberry Pi.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D242