Persist vulnerability results
Summary:

When vulnerabilities are discovered on a device, they are recorded in
the config tree.  We track when a vulnerability was discovered, when it
was last seen, and when we last reported it.  apcfg.GetVulnerabilities()
will extract this vulnerability state from the config tree and marshal
it into a Go struct for easy consumption.

For each client, record when our port and vulnerability scans start and
finish.

Test plan:

Manual with a client Pi vulnerable to CVE-2018-6789 and an XP VM
vulnerable to CVE-2018-6789 (Wannacry)

Differential Revision: https://ph0.b10e.net/D250