Makes DHCP lease info generally available
Summary:
We currently parse a device's DHCP lease in an attempt to determine whether to
run as a satellite node or not.  This change generalizes that lease parsing to
make more information available to a wider range of consumers.  The
functionality has been moved from ap_common/aputil to ap_common/network.

If the circus of goroutines involved in hostapd monitoring do not respond to a
shutdown instruction in a timely fashion, networkd will shoot itself in the
head.  This will force a complete networkd restart, which should ensure that
hostapd comes back online.  In addition, by using SIGABT, we will trigger a dump
of all goroutine stack traces for post-mortem debugging.

The vestigial aputil.RunAbort functionality has been removed.

Test Plan:
manual on a Pi, with explicit delays added to simulate hostapd
deadlocks.  Verified that a satellite node is still able to identify
itself as such.

Reviewers: cat, dp

Reviewed By: cat, dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D266