Version the inter-daemon RPC APIs
Summary:
Add Version fields to the MCP and ap.configd protobuf command and reply
structures.  Add Ping commands to both APIs that can be used to check for
daemon liveness and to verify that the two components are using the same
version of the API.

For ap.configd, the scope of the API version includes the config tree
structure and contents as well as the protobuf structure.

This change has limited utility on single-node installations, where all of
the components should be built together and installed from a single
package.  It is more useful in a multi-node installation, where each node
may be running from a different .deb package.

Test Plan:
manual regression testing.  Added new tests to configd_test.go
to test failing scenarios.

Reviewers: stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D284