Fix crash at startup in httpd
Summary:
This commonizes the code that ingests the DNS blocklist and makes `ap.httpd` use
it, which fixes T104.  This eliminates the last consumer of `data/phishtank`, so
that is removed.

Test Plan:
Install package on Pi, reboot, make sure that `ap.httpd` starts up, test that a
DNS lookup of `www.albinoblacksheep.com` gives us back the ring's address for
the AP, and that an HTTP request to that host sends us to the AP's phishing
page.

Reviewers: nils, stephen, dp

Reviewed By: nils, stephen, dp

Subscribers: buildbot

Maniphest Tasks: T104

Differential Revision: https://ph0.b10e.net/D301