Remove all references to blacklists and whitelists
Summary:
The terms "white" and "black" are often used to mean "good" and "bad", but that
association is racially insensitive.  Despite the common usage of whitelists and
blacklists, it's easy enough to change the terms to allowlists and blocklists.
In fact (for whatever reason), about half of the code already referred to
blocklists instead of blacklists, with the result that sometimes a variable
named blacklist would point to pathname which included the word blocklist.

This commit renames everything to use "allow" and "block".

Test Plan:
Having built modified versions of the tools to point to a different bucket, run
cl-aggregate to put the various lists into the cloud, and verify that all the
filenames use "allow"/"block".  Installing the same version on a Pi, and showed
that it downloaded the files as expected, into the new pathnames.

Reviewers: nils

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D305