Files written by cl-aggregate aren't really CSV files
Summary:
The biggest difference is that they don't have the same number of fields
on each line, which we correct here.  The other differences (CSV files
don't support comments, and usually require header lines) are easier to
work around with existing CSV parsing libraries, should we choose to use
them, so we simply document the discrepancies.

Test Plan:
Run `cl-aggregate` (modified to not upload) and ensure that the CSV files it
would upload have the same number of commas in each line.  Run the appliance
with both the old and the new format and see that `ap.dns4d` and `ap.httpd`
start up and perform their anti-phishing duties correctly.

Reviewers: nils

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D302