Improve multi-node robustness
Summary:
This is a significant overhaul of the internals of mcp, with the intent of
making the system more robust, primarily in the presence of node failures
but also for some daemon failures.  This also provides more visibility into
the status of a multi-node system.

Each daemon now has a dedicated goroutine that is responsible for managing
the entire lifecycle - not just when the daemon is actually running as a
child process.  Rather than explicitly launching/stopping daemons, ap-ctl
commands now just establish 'goal' states for one more more daemons.  Each
daemon's goroutine will take whichever action(s) are necessary to drive the
daemon to its intended state.

mcp daemons on satellite nodes exchange daemon status with the mcp daemon
running on the gateway.

When a satellite re-establishes contact with a gateway node, all running
daemons will restart.  This ensures that any config state that was changed
while disconnected is properly incorporated by the satellite daemons.

mcp's state is now included in the 'status all' output.

There is now an explicit 'blocked' state.

Daemons will restart if/when any of their dependencies restart.

While they mean well, daemons should not send an 'OFFLINE' update.  That
should only be set by mcp when the child process dies.

Test Plan: manual on a 2-node Pi config

Differential Revision: https://ph0.b10e.net/D297