Send the DHCP DISCOVER options to the log
Summary:
Have some identifiable DHCP information beyond the MAC address and offered IP
address in the log.  This allows easy access to the parameter request list in
particular, which is used for DHCP fingerprinting.  The list of requested
options and vendor class identifier may be useful there, too.  All this data is
published into the cloud, too, but that's harder to see.

Test Plan: Manual checking of the log.

Reviewers: nils, stephen, dp

Reviewed By: nils

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D322