Update timestamp on sysroot sum file
Summary:
The timestamp on $(SYSROOT)/.$(SYSROOT_SUM), which is a dependency for
anything doing a cross-build, comes from the timestamp in the tarball
it's part of.  Thus it's always going to be out of date with respect to
the tarball, and every build will re-unpack the tarball unnecessarily.

Test Plan: See that builds don't constantly show the tarball being unpacked

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D319