Upload stats and firewall data to the cloud
Summary:
Periodically upload the local files containing traffic statistics and
firewall logs to the cloud, then remove the local files.

This initial version requires google credentials to be configured on the
appliance to access cloud storage.  Eventually, the target 'signed URLs'
will be generated in the Brightgate cloud, allowing us to remove this
sensitive data from the appliance.

Buckets are now config properties, rather than being hardcoded in the
source.

Test Plan:
running on a Pi-based appliance.  Manually verified that the
storage bucket for the appliance was being populated.

Differential Revision: https://ph0.b10e.net/D331