Cloud data manipulation
Summary:
cl-dtool has been added to mananage and manipulate data in the cloud.  It
currently has 2 primary functions.

   - It can 'merge' multiple archives into one larger one.  This can be
       used to aggregate our 5-minute snapshots into full-day archives.

   - It can extract multiple datasets from the JSON archives into CSV
       files, suitable for bulk upload into BigQuery.

The data structures used for archiving watchd data have been moved into
common cloud/appliance code.

When snapshots are uploaded to the cloud they are assigned a Content-Type,
which is maintained as part of the object's metadata.  This field is
consumed by cl-dtool, which uses it to determine how the file should be
parsed.

ap-stats and the watchd api have been removed.  Now that the data is being
uploaded to the cloud, this on-appliance facility is no longer needed.

Test Plan:
Manual with office and home test machines.  Inspection of the
resulting datafiles by hand.  Verified that data generated on an
appliance can be landed in BigQuery tables.

Differential Revision: https://ph0.b10e.net/D341