JWT expiry isn't handled properly in ap.rpcd
Summary:
I botched the JWT expiry logic, because I thought that the gRPC client would
hang up and reconnect (which is totally wrong for multiple reasons).  The new
code moves the MakeGRPCContext() call to the point of sending the request, and
the credential will regenerate automatically as needed.

Test Plan:
- CI testing
- Longer running tests on Pi to validate expiry logic

Reviewers: duvall, stephen

Reviewed By: duvall

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D345