Add a global channel setting
Summary:
For inter-AP roaming to work, each AP needs to be using the same SSID
and channel.  We already have global settings for SSID and passphrase,
but the channel needs to be set for each NIC individually.  This change
adds a global per-band channel setting

Test Plan: currently running on a 2-node system

Differential Revision: https://ph0.b10e.net/D366